requirements.txt
aiohttp>=3.12.15
anyio>=4.9.0
bubus>=1.5.6
click>=8.1.8
InquirerPy>=0.3.4
rich>=14.0.0
google-api-core>=2.25.0
httpx>=0.28.1
portalocker>=2.7.0,<3.0.0
posthog>=3.7.0
psutil>=7.0.0
pydantic>=2.11.5
python-dotenv>=1.0.1
requests>=2.32.3
screeninfo>=0.8.1
typing-extensions>=4.12.2
uuid7>=0.1.0
authlib>=1.6.0
google-genai>=1.50.0,<2.0.0
openai>=2.7.2,<3.0.0
anthropic>=0.72.1,<1.0.0
groq>=0.30.0
ollama>=0.5.1
google-api-python-client>=2.174.0
google-auth>=2.40.3
google-auth-oauthlib>=1.2.2
mcp>=1.10.1
pypdf>=5.7.0
reportlab>=4.0.0
cdp-use>=1.4.4
pyotp>=2.9.0
pillow>=11.2.1
cloudpickle>=3.1.1
markdownify>=1.2.0
python-docx>=1.2.0
browser-use-sdk>=2.0.12
Flask
Flask-Cors
langchain

browser-use
langchain-google-genai
langchain-groq
langchain-openai
